ash. just ash.

+++ include/field_effect_helpers.h
@@ -28,7 +28,7 @@ void sub_8126BC4(u8 unk_1B, u8 r6, s16 x, s16 y);
 void sub_8127ED0(u8, u8);
 void sub_8127F28(u8, u8, s16);
 u8 sub_8128124(u8 id);
-void ash(s16, s16, u16, u8);
+void ash(s16, s16, u16, s16);
 void SetUpReflection(struct MapObject *mapObj, struct Sprite *sprite, u8 a);
 u32 oe_exec_and_other_stuff(u8 fieldEffectId, struct MapObject *mapObject);
 u8 sub_8126FF0(u8, u8, u8, s16, s16);
